requirements.txt
allure-pytest==2.9.45
allure-python-commons==2.9.45
attrs==21.4.0
certifi==2022.6.15
charset-normalizer==2.0.12
colorama==0.4.5
idna==3.3
iniconfig==1.1.1
jsonschema==4.6.0
packaging==21.3
pluggy==1.0.0
py==1.11.0
pyparsing==3.0.9
pyrsistent==0.18.1
pytest>=7.1.2
requests==2.28.0
six==1.16.0
tomli==2.0.1
urllib3==1.26.9
